requirements.txt
numpy
tqdm
spacy
django
bottle
tensorflow=1.15
djangorestframework
ujson